tome/requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
gunicorn==23.0.0
idna==3.8
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.2
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
pytz
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.20
